moduler.py 업데이트

+++ moduler.py
@@ -48,7 +48,7 @@ def load_stock_data(type, code) :
     tmp.columns = ['ETF코드', '종목명', '보유량', '평가금액', '비중']
     tmp['보유량'] = tmp['보유량'].astype(float)
     tmp['평가금액'] = tmp['평가금액'].astype(float)
-    tmp['비중'] = tmp['비중'].astype(float)
+    # tmp['비중'] = tmp['비중'].astype(float)
 
     return tmp
